AdVise/etl/db/requirements.txt
# Marketing pipeline (preprocessing, loaders) — same stack as in parent etl/requirements.txt subset
numpy>=2.0.0
pandas>=2.2.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
